Security overhaul
All lines read from a Gemfile are now scrubbed so as to ignore any lines
that do not begin with "gem ". As a side effect, we no longer collect
group or platform data for the gems, which isn't a big deal for the
site.

As a further security check, the $SAFE level is elevated from 0 to 3 in
the context of the instance_eval.